+++ cmd/ash/main.go
@@ -1,9 +1,8 @@
-// Command ash is the experimental standalone Go implementation of Ashley.
+// Command ash is the standalone Go implementation of Ashley.
 package main
 
 import (
 	"errors"
-	"fmt"
 	"github.com/LBYPatrick/ashley/internal/cli"
 	"github.com/LBYPatrick/ashley/internal/execution"
 	"os"
@@ -11,7 +10,7 @@ import (
 
 func main() {
 	if err := cli.Run(os.Args[1:], os.Stdout, os.Stderr); err != nil {
-		fmt.Fprintln(os.Stderr, "Error:", err)
+		cli.WriteError(os.Stderr, err)
 		var exit execution.ExitError
 		if errors.As(err, &exit) {
 			os.Exit(exit.Code)